Html, CSS and Javascript changes to make a grid
I tweaked the html to make a button that lets me create a new button that I will later use as a reset button. The css was used to make a thick border around the grid I made and give it a solid border. The Javasctipt file was changed to simplify how big I wanted my grid and to make individual divs that the css would display as boxes or as a grid. The other code is a placeholder code to prepare for the code to make the reset button reset the scene.