html element name changes, css styles for button and Javascript code change to enable prompts.
In html I changed the id from reset button to changeSize so I can clearily see how and where to use it in my javascript code. In css I changed some of it to make the grid have the capabilities to be larger but limit it and chnaging the button to make it round. In javasctipt I created my button event listener that will display a confirm message and then a prompt to set the size of the grid.