Html and Javascript heading , prompt and resizing changes
I changed the heading on the html file and changed the button id to have it more specific in my code. In javascript I rearranged and code such as the id and class events and event listeners and appended them so that my button can do the job of resizing my grid when button is pressed then a pompt will come up and the user will be able to select a number from 1 to 100 and no larger also it will have to be a number or it will prompt the user to input the correct values. The comments from my old code were left in to refresh me on how to use event listeners and how to append them. I also incorporated new properties into the code that helped me to determine how the code would change the grid size .